+++ models/base.py
@@ -1,4 +1,3 @@
-from sqlalchemy.orm import DeclarativeBase
+from db.models.base import Base
 
-class Base(DeclarativeBase):
-    pass
+__all__ = ["Base"]